+++ src/main/java/net/tassia/pancake/plugins/cli/CommandInfo.kt
@@ -1,4 +1,4 @@
-package net.tassia.pancake.cli
+package net.tassia.pancake.plugins.cli
 
 /**
  * Holds information about a command.
