+++ src/main/java/net/tassia/pancake/plugins/cli/command/HelpCommand.kt
@@ -1,7 +1,7 @@
-package net.tassia.pancake.cli.command
+package net.tassia.pancake.plugins.cli.command
 
-import net.tassia.pancake.cli.Command
-import net.tassia.pancake.cli.CommandInfo
+import net.tassia.pancake.plugins.cli.Command
+import net.tassia.pancake.plugins.cli.CommandInfo
 
 /**
  * The 'HELP' command. Shows helpful information.
